Turned drivers-maker into a Dropwizard service
This commit implements drivers-maker as a dropwizard service, providing
an API to utilities to generate the benchmark packet to be sent to
Faban.

In particular, this commit implements:
- POST /config, which receives a benchflow YAML configuration and
converts it to a Faban compliant XML configuration. the fields javaHome
and javaOpts must be configured through Dropwizard configuration file.
- Several upgrades to the config-converter module, including updated
pom.xml files, some additions to the API of the converter objects, and
the elimination of warnings in the compilation phase.