Adds serialization of experiment definition model to YAML
This commit implements the serialization of the  experiment definition
model to YAML; this means that now it’s possible to parse a YAML
`benchflow-benchmark` file to retrieve the model, and dump it to YAML
again. This will be useful when we will have to programmatically change
the model and save the changes.